Move code generation out into a module and run snapshot tests
