+++ unit-test/config.json
@@ -2,7 +2,8 @@
   "spec_dir": "unit-test",
   "spec_files": [
     "**/*[sS]pec.?(m)js",
-    "**/*.js"
+    "**/*.js",
+    "!unit-test/script-overload-snapshots/out/*.js"
   ],
   "jsLoader": "import",
   "helpers": [
