+++ unit-test/script-overload-snapshots/config/2.json
@@ -1,0 +1,6 @@
+{
+    "single": {
+        "type": "string",
+        "value": "meep"
+    }
+}
